add logic to update connections for MSI
